EV-5735: Don't include Windows.h in headers if we can avoid it
ofmem.h is included transitively any time we include a DCMTK header, and
if that includes Windows.h, we end up paying for that many times,
rendering it the single most expensive header in all of EasyViz.

So just don't do it. We only need a couple of Windows functions, and
they can be forward declared just fine.

A better long-term solution would be to enable C++11/C++17 when
building DCMTK, but that requires a bit more work, as the DCMTK team do not
appear to have tested this on Windows. By default, MSVC sets __cplusplus
to its C++98 value even in C++17 mode, and DCMTK uses this
several places to test the C++ version.