EV-9071: Ensure that decompressed format matches what is in the dataset
This commit corrects per-frame decompression when the compressed
stream produces 8bit data but the target dataset has
bits-allocated = 16.

Full dataset decompression handles this by changing BitsAllocated in the
dataset but that is not an option when doing per-frame decompression.

Instead we must ensure that the decompressed pixels match the dataset.

This is a long-existing bug in dcmtk but it hasn't bothered us until
recently because of a bug in DICOMImagePixel

At this time the problem hasn't been reported upstream