EV-9563: Implement full support for per-frame decompression (RLE codec)
The existing RLE codec would use the deprecated DcmCodec::determineStartFragment
to find the start fragment for a frame. Clearly that is not going to work
as that only supports jpeg.

change it to a call to DcmCodecList::determineStartFragment and implement
DcmRLECodecDecoder::findStartFragment to ensure that it works even if
the fragment offset item is absent.