EV-11595: Fix memory leak in dcmitem::insert
DcmItem::insert(elem, true) asks to replace any existing value of
elem.getTag with elem. This fix ensures that the existing value
is deleted when elem is inserted.

Note that the fix is in DcmElementList::Impl::insert() which is our
own addition to dcmtk and therefore this issue only exists in our fork.