EV-11944: Don't use a global mutex when accessing configuration values
In some cases, when building MINT datasets we spend almost 80% of the
time waiting for a mutex around a configuration variable.

This is a two-part fix. In general, we benefit from replacing the mutex
with a read-write lock, allowing concurrent readers.

However, when benchmarking this fix it still shows significant time
spent waiting on this lock (now down to perhaps 50% of the time).

So to further improve on this, create a specialization for booleans,
which drops the lock and uses a std::atomic.

The reason we don't always use atomic is that the OFGlobal class is
also instantiated with types that cannot be made atomic<> (strings, for
example).

We *could* make a more comprehensive fix, using type traits and
metaprogramming to apply the atomic<> specialization for all types where
it's legal, but let's keep it simple. Only the boolean specialization
has so far shown itself to be a performance problem.