EV-11944: Make isspace() faster when processing UID strings
The C standard isspace() function looks up the current locale (because
tht might in theory affect which characters constitute whitespace),
which becomes a performance bottleneck when called often.

In benchmarking, after fixing the configuration mutex in the previous
commit, most of the time building MINT datasets is now spent by
DcmUniqueIdentifier::makeMachineByteString, in the call to isspace().

For UID data, we're not going to see exotic locale-dependent whitespace
characters, so use a simpler reimplementation of isspace().