Remove OFGlobal::xget
The function is unused, and it is not thread-safe, even when compiled
with WITH_THREADS.

The function returns a reference to the global data, and while it locks
a mutex during the call, the reference will be accessed after the
function returns, when there is no mutex protecting the value.

As the function cannot be made safe, and is unused, let's just remove
it.